Move action view to image pager activity
